+++ src/modules/Opportunities/components/registration-field-view/texts.php
@@ -1,0 +1,8 @@
+<?php
+
+use MapasCulturais\i;
+
+return [
+    'Endereço completo' => i::__('Endereço completo'),
+    'Etapa' => i::__('Etapa'),
+];